Заполнил NavigationDrawer. Добавил активити "Настройки" и "О программе".
