+++ app/src/main/res/drawable/side_nav_bar.xml
@@ -2,8 +2,8 @@
     android:shape="rectangle">
     <gradient
         android:angle="135"
-        android:centerColor="#4CAF50"
-        android:endColor="#2E7D32"
-        android:startColor="#81C784"
+        android:centerColor="#afad4c"
+        android:endColor="#7d7d2e"
+        android:startColor="#ead12e"
         android:type="linear" />
 </shape>
