+++ app/src/main/res/layout/activity_about.xml
@@ -4,15 +4,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.exwhythat.mobilization.ui.activity.MainActivity"
-    tools:showIn="@layout/app_bar_main">
+    tools:context="com.exwhythat.mobilization.ui.activity.AboutActivity">
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Hello World!"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        android:layout_gravity="center"
+        android:text="@string/action_about"/>
+
 </FrameLayout>
